# src/docs/swagger.yaml
openapi: 3.0.0
info:
  title: URL Shortener API
  description: Advanced URL Shortener with Analytics and Rate Limiting
  version: 1.0.0
servers:
  - url: https://urlshort.shadowdragon.dev
    description: Production server
  - url: http://localhost:3000
    description: Development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
paths:
...
  /auth/current-auth:
    get:
      summary: Get current authentication data
      tags:
        - Authentication
      responses:
        200:
          description: Current authentication data
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                      email:
                        type: string
                      name:
                        type: string
                      avatar:
                        type: string
        401:
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    get:
      summary: Logout user and invalidate current token
      tags:
        - Authentication
      security:
        - BearerAuth: []
      description: Requires a valid access token. The provided token will be blacklisted and can no longer be used.
      responses:
        200:
          description: Successfully logged out
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Logged out successfully"
        401:
          description: Unauthorized - Token missing or invalid
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Access token required"
                  message:
                    type: string
                    example: "Please provide a valid access token"
        403:
          description: Forbidden - Token already revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Token has been revoked"
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Logout Error"
                  message:
                    type: string
                    example: "Failed to logout"
...